Create Slider component for requiredUpvotes and xpPoints
